checkpatch: Fix unreasonable warning on author's sign-off.
This patch fixes the warning (WARNING: Missing Signed-off-by: line by nominal patch author '' ) asserted when executing the checkpatch.pl.
As we create a commit we can see in  (git show 'some hash') that for the author of the commit there is no "From" part. 

Changed the regexp that looks for the author to "Author" before it was "From".